Use LONGBLOB column type for lcache values
It's common for `options:alloptions` (and potentially other cache items) to be a lot bigger than 64kb. Currently this can result in a `null` being stored, which transports us into a hellscape of database thrashery. An incomplete serialized object goes into the `lcache_events` table due to the size constraint, and then fails to unserialize when being read, and then we're fail-boating.

We will also guard against this in LCache core - but we must make the schema ready to store much larger values.